Updates LibTomCrypt from version 1.16 to version somewhere between 1.17 and 1.18.
On or about October 16, 2007, I received a tarball from Tom St. Dennis
ostensibbly containing the 'most recent' version of LibTomCrypt. This
version was apparently more recent than the tarball for version 1.17
on libtom.org. This commit brings the code for LibTomCrypt in
depends/libtomcrypt up to speed with this tarball.