Set order of memory de-allocs to last-alloc'ed, first-freed.
In scanning the history of this project, I note that Mitch Bradley
painstakingly set the order of memory-freeing to be the reverse of the
order in which memory is allocated. Apparently, this allows simpler
memory management. In keeping with this, I've set the order of
memory-freeing in ecc_validate_key() to be the same: reverse of the
order of memory allocation.