Changed rsa_make_key() calls to use new function-signature.
Previous patches to LTC changed the function-signature for
rsa_make_key(). In keeping with ANSI standards, this function now
takes in two different sources of randomness. That same standard also
specifies that if the randomness-sources are PRNGs, then they should
be seeded with different random seeds. It looks like the tools makekey
and cli_tool both simply use the system rng, so I changed the code to
pass that same source in twice.

(Note: untested!)