Build: can now build with libtfm.a and libtomcrypto.a from rpms
So users can build it having installed tomsfastmath-audited-devel and
libtomcrypto-audited-devel which contain the .a files.

The 'depends' dir stays, and can still be used; we don't want to break
WMBs workflow. We don't include 'depends' however when creating a
tarball (see .gitignore).

The shlib builds are removed & deprecated -- the rpms provide shlibs
just fine.