zhashfs: detect fill pattern and suppress blocks in output
Allows the unused filesystem blocks of the image to be left out of the
.zd file, resulting in a factor of two reduction in fs-update times on
a typical build on a 4GB microSD.  Greater reduction occurs on larger
microSD or SD cards.

Has no effect if a fill pattern file is not found.

from olpc-os-builder a17b54443ecceea57df5dfc05145da2d085e2fba